requirements.txt
flask
shopifyAPI
python-dotenv
